% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darts_pipeline.R
\name{Darts_replicate}
\alias{Darts_replicate}
\title{The Darts Bayesian Hypothesis Testing function with replicate}
\usage{
Darts_replicate(
  in_fn,
  out_fn,
  iter = TRUE,
  rescale_meth = 1,
  C = 0.05,
  rho_fn = NA,
  estim_groupVar_prior = FALSE,
  is_paired = FALSE,
  pooling = FALSE,
  epsilon = 0.01,
  verbose = 1,
  random_state = 777,
  thread = NULL
)
}
\arguments{
\item{in_fn}{Input file name. Required. The input file is in rMATs output format, requiring columns ID,I1,S1,I2,S2,inc_len,skp_len. Additional columns are allowed. Multiple replicates should be separated by comma within columns I1,S1,I2,S2.}

\item{out_fn}{Output file name. Required. This function will directly write results to the file.}

\item{iter}{Use iteration or not. Logical, TRUE or FALSE. Default TRUE if no informative prior provided. If TRUE, it will start from uninformative prior (if no informative prior provided), and update the prior based on the posterior from the last round, until convergence. Note that if informative prior is provided, it will be set to FALSE with a WARNING message, even if the user specificed TRUE.}

\item{rescale_meth}{Rescale method. It will only be considered if informative prior file is provided, taking value 1 or 0. If 1, it will rescale by binary prediction, and by bias otherwise. Default 1.}

\item{C}{The cutoff to define significant difference, i.e., delta psi values, where psi is the exon inclusion rate. Default 0.05.}

\item{rho_fn}{Informative prior file name. The file format is from Darts_DNN output. If not provided, it will start from uninformative prior. We suggest setting iter=TRUE if no informative prior.}

\item{estim_groupVar_prior}{If output plot for estimated group variance prior. Default FALSE. If TRUE, if will output a plot with name groupVar.pdf in the working directory.}

\item{is_paired}{If data are paired between groups. Logical, TRUE or FALSE. Default FALSE.}

\item{pooling}{If pooling different replicates together as a single input for inference. Logical, TRUE or FALSE. Default FALSE. Suggest setting to FALSE unless necessary.}

\item{epsilon}{Convergence criteria. It will only on considered if iter=TRUE. Default 0.01.}

\item{verbose}{Default 1.}

\item{random_state}{random seed. Default 777.}

\item{thread}{Number of thread to perform parallel computing. Default will use all available threads if not provided.}
}
\value{
A dataframe of posterior probability. The following columns will be added appending to the original input file.
\item{rho}{The final prior P1, proportion of differential splicing events, from the last iteration to generate this output.}
\item{psi1}{Exon inclusion rate (psi values) for the first group. Multiple replicates will be separated by comma.}
\item{psi2}{Exon inclusion rate (psi values) for the second group. Multiple replicates will be separated by comma.}
\item{mu.mle}{Maximum likelihood estimator of mean psi value for group 1.}
\item{delta.mle}{Maximum likelihood estimator of mean difference in psi values in group 2 compared to group 1.}
\item{post_pr}{Posterior probability of this event being differential splicing. We recommend using post_pr > 0.9 to define significant differential splicing, post_pr < 0.2 to define no differential splicing, and in between as unsure.}
\item{covg}{Mean coverage for this event.}
}
\description{
It calculates posterior probability of differential splicing, taking input files from rMATs output (or similar file formats). It also provides an option to use iterative empirical Bayesian approach to update the original prior.
}
\examples{

out = Darts_replicate(in_fn = "test_rmats_rep.txt", out_fn = "test_darts_rep.txt", iter=T)

}
